+++ src/sources/lol_game_report_riot_api.hpp
@@ -3,6 +3,7 @@
 #include "lol_game_report_types.hpp"
 
 #include <QObject>
+#include <QJsonObject>
 
 #include <functional>
 
@@ -14,9 +15,11 @@ class riot_api final : public QObject {
 public:
 	explicit riot_api(QObject *parent = nullptr);
 	void enrich(report value, std::function<void(report, QString)> complete);
+	void set_diagnostics(std::function<void(const QJsonObject &)> callback) { diagnostics_ = std::move(callback); }
 
 private:
 	QNetworkAccessManager *manager_{};
+	std::function<void(const QJsonObject &)> diagnostics_;
 };
 
 } // namespace sources::lol_game_report